Completed simple game showing what Phaser can do. NOTES:  - BasicGame.Game code needs to be cleaned up  - As well as index.html and Boot.js  - Game.js could be redone in a better way, possibly using if-statements to check if a stage has been completed or not, then choosing the appropriate image/animation to show.
Issues:
- Had an issue trying to increment levelCounter. levelCounter is what will
be used to count if a level is completed or not. Once all levels have been
completed (in this case, when levelCounter is 2), the game will start the
game over screen.  I was unable to get this to work so I used a small if-
statement.  I hate if statements when comparing multiple bools.  >_>